requirements.txt
aiogram
loguru
openai
Pillow
openai_async